requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.1
python-dotenv>=1.0.0
requests>=2.31.0
pydub>=0.25.1
speechrecognition>=3.10.0
gtts>=2.4.0
pandas>=2.0.0
numpy>=1.21.0
python-docx>=1.1.0
Pillow>=10.0.0
audio-recorder-streamlit>=0.0.10
mistralai>=0.4.0
streamlit-javascript>=0.1.5
reportlab>=4.0.0
fpdf2>=2.7.0
